chore(Components): refactor List component files
